requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib==1.7.4
bcrypt==4.0.1
python-dotenv
pydantic[email]
requests
groq